expander, move: Move expander package to state
When using the expander we really need to specify that it's a state
expander so it make sense to rename the package as "state" and the
constructor as "NewExpander" so we can do "stateExpander := state.NewExpander".
